refactor(prisma): update schema to include User model and Inventario relation
This commit modifies the Prisma schema to replace the Cliente model with a more comprehensive User model, adding fields like password and role. Additionally, the Libro model now includes a relation to the new Inventario model to manage inventory URLs. These changes improve the schema's structure and functionality to better align with the application's requirements.